requirements.txt
tqdm
psycopg2
pillow
numpy
